Add files via upload
James is a frog who is currently sitting on a rock. The rocks around him are arranged in a grid of size
A × B, with co-ordinates ranging from (0, 0) to (A − 1, B − 1). James is sitting at (Jx, Jy). James can
jump on adjacent rocks which are to the North, East, West, or South of its current rock. Some rocks are
unstable and hence he cannot land over those rocks. A fly is sitting on a rock which is at the coordinate
(Fx, Fy). James has to reach the rock on which the fly is sitting (it is assured that the rock on which James
/ fly is sitting is stable), by jumping on only the stable rocks. Is it possible?